requirements.txt
python==3.10.13
numpy==1.26.2
pandas==2.1.4
tqdm==4.66.1
biopython==1.81
lingpy==2.6.13
matplotlib==3.8.2
ete3==3.1.3
scipy==1.11.4
regex==2023.10.3
